CLOB to SQL Script
This is a simple function that makes use of the APEX_STRING API to take any CLOB, and transform it into a SQL script. The resulting SQL script will reconstruct the CLOB and call your procedure to process it. This is handy for exporting data (e.g. JSON documents) from one system and incorporating it into an installation script to be imported elsewhere.